made some python3 changes moved the error class out to it's own file
