+++ pyairwatch/system/admins.py
@@ -1,5 +1,6 @@
 from .system import System
 
+
 class Admins(System):
 
     def __init__(self, client):
@@ -20,4 +21,12 @@ class Admins(System):
             role={role}
         """
         response = System._get(self, path='/admins/search', params=kwargs)
+        return response
+
+    def create_adim_v1(self, user_data):
+        """
+        Performs necessary checks and Create a new basic Admin user.
+        """
+        path = '/admins/addadminuser'
+        response = System._post(self, path=path, data=user_data)
         return response
